[Issues-1] Generation of recursive data fails
[Problem]

Generation of recursive data structure fails because of the stack overflow.
Structure that has a property with pointer to itself is one of the examples.

[Solution]
The issue is that generator.Ptr(arb Arbitrary) will enter in a recursive
loop as it tries to create Generator for arb. Creation of arn Generator
is initated whether the pointer should nil or not.

In order to avoid this issue, randomness of arb's generator creation
should be done before Generator for Ptr is created. In order to do so
Arbitrary signature needs to be changes, as Arbitrary needs to have an
access to *rand.Rand to implement randomness.

Arbitrary signature is now: type Arbitrary(target reflect.Type, r *rand.Rand) (Generator, error)

This also allows a implementation of generator.OneOf. Two other arbitraries
in generator package have been added: generator.PtrInvalid, generator.PtrValid.

Closes #1