[feature/issues/8] Hash input of generated function to create seed for Random
[Problem]

generator.Func() should create pure functions, for same input a same output
should be generated. 2 generated functions with same signatures should
be different from one another, such that for same input they generate different
output.

[Solution]

Function input is encoded to it's string representation and than hashed
using fnv.Hash64. Input is mapped to int64, so it can be used as a seed
for Random of function output generators.

Added a way to create string representation for all Go's basic types except
interfaces{}. This is done in arbitrary package for now, with function
EncodeToString(). Idea is for it to be used by check package as well.

There was an issue with previous implementation of passing Random between
Generators. All Generators were splitting from parent Random, but this
proved to be a bad strategy because these generators couldn't be used
for generating function output predictably. As soon as the Random is split
seed altered in generated function body didn't affect the generators for output,
because they were using completely different Random. Random is now split only
when it's seed needs to be manipulated for predictable output.

Close #8