[bugfix/issues/16] Generation of composite structures fails
[Problem]

```go
type C1 struct {}
type C2 struct {}

type C struct {
    C1
    C2
}
```

If there is a type as shown in the example above, assignment of generated
value to type C will fail.

[Solution]

The issue is with how creation of Struct is done by reflect package. In
the current itteration target type is constructed from struct fields,
which in this case is:

```go
struct { C1: struct{}, C2: struct{}}
```

Embedded types have a field in a parent structure that nmatches the name
of a type. But it seems that creating a structure from these fields is not
the same as embedding types into structure itself. Solution to this problem
is use target as type refernce during the call to New instead of using
struct fields to create structure's type by using reflect.StructOf.

As a side effect this change removes alot of code as well.

Close #16