[bugfix/issues/32] Shrinking of negative int64 fails to shrink around 0
[Problem]
https://github.com/steffnova/go-check/issues/32#issuecomment-899030288

[Solution]
Splitting shrinker in composition of 3 shrinkers:
  - Shrinker for negative numbers
  - Shrinker for positive numbers
  - Shrinker for 0

Shrinking of negative numbers is selected if N (target that needs to be shrunk)
is negative. The value will be shrunk towards 0 or limits.Max whichever is
lower.

Shrinking of positive numbers is selected if N (target that needs to be shrunk)
is positive. The value will be shrunk towards 0 or limits.Min whichever is
higher.

Shrinking of 0 returns 0, without shrinker as it is the converging point
for shrinking of both positive and negative numbers

Close #32