[feature/issues/54] Improve map shrinker and generator
[Problem]

Map generator should be able to generate a Map within provided size limits
if pool from which key values are generated has number of unique values
within the size limits. Time that it takes can be directly impacted with
the pool of unique values available for keys. In other words collisions
of generated key values should be taken into the account, without taking
a consideration of time required resolving these collisions.
It should also be possible to create maps with fixed size (by using same
value for size's max and min value).

Map shrinker should never shrink map size outside of size limits (which
is not currently the case). The main problem is that during shrinking of
keys, there could be a a key collision in which case map size would be
reduced (because of the multiple elements with the same key).

[Solution]

1. When map generator is generating key values, it takes into an account
all current map element's keys. If generated key is duplicate, key will
be regenerated until a unique value is created.

2. Map shrinker has been refactored. It is now built as a composition of
Slice and Array shrinkers. Map elements are represented as a slice of pairs
(key and value).

3. FilterPredicate function is added to arbitrary package. FilterPredicate
allows creation of arbitrary predicate functions in a runtime using reflect
package.

4. Filter shrinker method is added. It uses predicate function to filter
out shrink values that satisfy it. Filter method for generator.Arbitrary
now uses Shrinker.Filter (previously it didn't have shrinker).

Close #54