[bugfix/issues/72] Improve bias for generated integer values for symmetric range
[Problem]

Problem is described here: #72

[Solution]

In order to perserve symmetry ratio for generated numbers (50:50 ration between
negative and positive numbers when range is [-n, n]), 0 is considered a positive
number and negative numbers start from -1. This way negative numbers can be
represented with same number of bits as positive, and will ensure that range of
positive and negative values is expanded at the same rate.