[bugfix/issues/76] shrinker.Bind is causing panic
[Problem]

Problem is described here: #76

[Solution]

The issue is that the last failing shrinker is not persisted during the
shrinking process. Last failing shrinker should be the returned by
binder and it's shrink must falsify the property.

The second problem is with the shrinker.Retry as it wasn't working as
intended. This has been resolved as well