Shrinkers should accept the value they are shrinking
[Problem]
Shrinker enclosure value they are shrinking, which doesn't allow to use
same shrinker for different value. This functionality is neccessary to be
able to chain multiple shrinker's together (one example is chaining
shrinker with different strategy of shrinking slice elements)

[Solution]
When passing values to shrinkers there must be a context on what that
value represents. If value is generated through combinators, that value
will pass through multiple shrinker combinators and as a result, final
value passed from generator cannot be passed through all combinators.

In case of Uint64->Uint16 mapping, final value will be uint16, but underlaying
generator that was used was Uint64. Passing uint16 down to combinator chain
will result in a error from map shrinker as mapper input doesn't match the
value passed to Map combinator.

There are 2 possible solutions to this problem.
  1. Mapper accepts, mapper for mapping from Uint64->Uint16, and another
  mapper that can "unmap" the Uint16->Uint64. This would add complexity to
  API, as now two functions need to be supplied to Generator when map
  combinator is used. The other problem is that, not all mappings are
  symetric, (they can't have unmapper)
  2. Value will contain tree like data structure where first node will be
  a final value, and child nodes will represent values from which final was
  constructed (recursevily). Every value will have historic data in itself.
  The API will in this case stay the same.

2nd approach has been taken as a solution. New type Arbitrary has been added
to arbitrary package. It has a Value that is reflect.Value and represents
a final value that will be passed to property. It has Elements property, which in
case of collection like structures (struct, slice, array, map) will have
Arbitrary representation for each container element. Precursors are arbitraries
that were used to generate parent arbitrary. In case of Map and Bind combinators
there will be only one Precursor.

Codebase has been updated to aclimate to this major change.

[NOTE]

- Generators ConstantFrom, OneFrom have been fixed to work properly
- OneFromWeighted is a new generator that allows selection of one of the
  generators based on weight (OneFrom is a OneFromWeighted where every
  generator has the same weight)
- Int and Float generators have been changed to be derivatives of Uint
  generator (they inherit shrinking from Uint with this change)
- 3 new shrinkers have been added:
    CollectionSize shrinks the size of collections (map, slice, array)
    CollectionElement shrinks collection elements one by one
    CollectionElements shrinks collection element all at the same time
- Map, Array, Struct and Slice shrinker and generators have been refactored