🐞 [bugfix/issues/80] Shrinking for Filter combinator not working
[Problem]
Shrinking process breakes when using Filter combinator

[Solution]
Shrinker combinator for Filter, on predicate failure wasn't passing
the last shrink value, but the last value that caused property to fail.
This caused undefined behaviour for shrinking process, thus breaking it

[Note]
Removed error for Map shrinker, number of Arbitrary elements didn't match
the size of the map. This could happen if there are two arbitrary elements
with the same key.