[bugfix/issues/86] Float32 and Float64 are generating values out of range
[Problem]
The issue is present only with negative range of values.
Error in conversion float32/float64 negative numbers to uint32/uint64.

[Solution]
There is no need to converting negative float numbers to positive. Bit
representation of float32/float64 is ordered for negative values. The
special care needs to be done for 0 as it has two values (+0, and -0).

For negative range -0 must be used. This change also removes requirements
for mapper uint -> float for both negative and positive values.

Close #86