[feature/issues/90] Move random implementation to generator package
[Problem]
It makes more sense for implementation to be in the generator package
as it is primarly used by it.

[Notes]
Additionaly Random no longer has Int64 method. This requirement was only
needed when dealing with length constraints which is int64 type

Added check for length constraints in chan, map and slice generators

Close #90